mysite/requirements.txt
django-filter>=2.2.0
djangorestframework>=3.10.3
Markdown>=3.0.1
social-auth-app-django>=3.1.0
social-auth-core>=3.3.3
django-extensions>=2.2.5